more cleanup moved files to appropriate locations removed unused components moved components to views
