--- src/components/tableField.vue
@@ -1,11 +1,0 @@
-<template>
-  <td>
-    <slot></slot>
-  </td>
-</template>
-
-<style lang="scss" scoped>
-td {
-  padding: 1em;
-}
-</style>
