+++ src/router/index.js
@@ -2,9 +2,9 @@ import { createRouter, createWebHashHistory } from "vue-router";
 import Pending from "../views/Pending.vue";
 import Records from "../views/Records.vue";
 import addNew from "../views/addNew.vue";
-import finalize from "../views/finalize/finalize.vue";
-import Report from "../views/report/Report.vue";
-import Invoice from "../views/invoice/Invoice.vue";
+import finalize from "../views/finalize.vue";
+import Report from "../views/Report.vue";
+import Invoice from "../views/Invoice.vue";
 import Settings from "../views/Settings.vue";
 import Summary from "../views/Summary.vue";
 import syncQueue from "../views/syncQueue.vue";
